Update for 1.1.2 version
- add python 3.6 to travis
- update version number
- update zstd from upstream to v1.1.2
- simplify tests
