+++ .travis.yml
@@ -6,6 +6,7 @@ python:
   - 3.3
   - 3.4
   - 3.5
+  - 3.6
 
 addons:
   apt:
@@ -21,7 +22,7 @@ install:
 script:
 - python setup.py test
 - python setup.py clean -a
-- python setup.py --legacy test
+- env LEGACY=1 python setup.py --legacy test
 - python setup.py clean -a
 - python setup.py --external test
 - python setup.py clean -a
